Issue 5402/2: Arrange for System::get_bound to return a Paper_column
Enforce that only Paper_columns may be set as bounds of Systems.
Whether an Item and a Spanner can be linked depends on the specific
type of each, so we use two virtual function calls to simulate
multiple dispatch.

This is part of a campaign to preserve more type information in the
C++ code, with the goal of reducing the amount of run-time type
checking and conditional code.