Issue 5340: Simplify Spacing_spanner::get_columns
The old algorithm filtered the system's list of columns into a list of
used columns, then searched the filtered list for the endpoints of the
range to return.

The search key was the "rank" of a column, which is its index in the
unfiltered list of columns (see System::add_column).  There is
therefore no need to search.

The new algorithm selects the desired range in the unfiltered list and
then filters that.

This commit also removes the unused function Paper_column::compare.