Issue 5346: Fix GC issue in skyline rotations
This was issue 5341 initially and is a fix for a problem introduced
during issue 5319.  The problem basically is that a rotated Stencil is
created and its stencil expression is getting stored on the heap
without retaining any pointer to it in areas scanned during garbage
collections.

This fix is a minimal one that merely avoids creating a stencil and
instead simulates the effect on the transform matrix to be expected
from using a rotated stencil.