Issue 5356: Fix result type of composed partial functions
When using \etc for composing multiple partial functions into a single
function, the resulting signature takes the trailing argument list
from the last such composed function.  However, the type of the
resulting function (music or event or scheme function) is that from
the first function.

The previous implementation got this wrong.