Issue 5338 - Fix 'out of sync' ps script defs
Commit 12fe78825798191ecb7e5a4ee3064679773ae1ab
broke encodingdefs.ps.

The glyphs are now included in
the same order as they are in
the emmentaler fonts.

The order of the glyphs does not
matter as long as the encoding
vector and the corresponding
commands match.

The change occured because I'd
decided to use otfinfo and sed
to generate the new code -
it seemed easier, safer and faster
than correcting the old code by hand.