Issue 5366: Move warnings out of find/create context functions
The motivation for this is that Context::find_create_context () and
find_context_near () should probably be merged for maintainability,
but one of the differences between them that must be dealt with is
that find_create_context () logs when it fails and find_context_near
() does not.  Adding warnings to find_context_near () risks being too
noisy, leaving the option taken here.

The new method Context::diagnostic_id (name, id) returns a formatted
string (e.g. "Voice" or "Voice = mel") for use in a log message.  It
is used for the warnings that are being moved as well as a few other
existing warnings to increase consistency.