Issue 5381/4: Change intermediate PDF filename for LaTeX
There was a possibility that intermediate PDFs are output as final PDFs
since the intermediate filename is the same as the final filename.

This commit changes LaTeX output directory to a temporary directory.

After generating the intermediate PDF,
we remove the temporary directory.
If LaTeX terminates in error,
make terminates before removing the temporary directory
and we can investigate LaTeX logs.