Issue 5381/5: Change intermediate PDF filename for DocBook
There was a possibility that intermediate PDFs are output as final PDFs
since the intermediate filename is the same as the final filename.

This commit makes change intermediate PDF filename for DocBook.

By default dblatex uses temporary directory
instead of current directory for build.
Therefore, only by changing the output filename,
the intermediate PDF is not generated as final PDF filename.