Issue 5391/2: Don't set origin on copied music explicitly
There would be some mild incentive if the origin were accessed an
inordinary amount of times (and thus leaving it unset would maximize
access times to it) but there is no evidence for that.  One reason
might be to ensure greater structural similarity between copies of
music, but Prob::equal_p has been changed to be robust against
differences here.