Limit angle of broken hairpins
Broken parts of hairpins don't look good when they are too short.
This is because their height at the broken side is hard-coded to
2/3 and 1/3 of the hairpin's height. So when the sibling's width is small
the ratio/angle is too high to be visually pleasing.

This commit (initial thoughts) limits the width/height ratio of broken
hairpins to 1/2.

It would be good to provide both the broken-heights and the ratio
as native grob properties for Hairpin, though.