Functions to invert chords or drop/rise chord notes
This feature was suggested and improved upon by
Davide Bonetti, with some help from David K:
https://lists.gnu.org/archive/html/lilypond-user/2019-01/msg00360.html

Included here are three new commands: \dropNote, \riseNote and
\invertChords, and a new Scheme function, move-chord-note.
The first two commands take an integer that selects which
note will be transposed one octave lower or higher within
every chord in the music expression that follows, thus allowing
for more sophisticated voicings (e.g. `drop 2’ and the like).
If the argument is out of range (if there aren’t enough notes
in the chord), it’s silently ignored and the chord’s left unchanged.

\invertChords accepts both positive and negative integers, and
inverts all chords accordingly (as in switching the root note’s
position, not as in reversing intervals like \inversion does).

These commands work with both explicitely-written chords and
\chordmode-entered chords (only when printed on a staff:
ChordNames are to remain unchanged).