Add glib-2.0 and gobject-2.0 library dependency
For a long time we relied
on pango/pangoft2 to pull
in the glib and gobject
libraries. That worked as
long as every distribution
used a pangoft2.pc file
that included both libs in
the 'Required:' line.

According to the pkg-config
documentation that was
definitely wrong, and the
problem was corrected on
opensuse tumbleweed
recently.

As a consequence building
lilypond failed with a
linker error as we use
symbols from both libs
directly in our c++ source
code.

Signed-off-by: Knut Petersen <knut_petersen@t-online.de>
(cherry picked from commit bc4f5fdc299ff21d0ee2de4462af30b22e1321c1)