Add tex/txi-pt.tex
We need txi-*.tex files for all the languages we
support. Without them, building of lilypond fails.

If you did not experience that problem, you were
lucky and texinfo is installed on your build system.
In that case the system's txi-pt.tex is used, see
Documentation/pt/web.texi2pdf.log.

txi-pt.tex is taken from texinfo 6.1 (we use that
version in gub).