ly: hel-arabic - comment out Sahnaz entry
In oriental, Arabic music sometimes different makam names correspond
to a makam mode. We must keep only the real makams

After a thorough search, I noticed that the makam Sahnaz that is in D
is in fact a transposition of the makam alhizazkar which is in C.

e.g.

\transpose c' d' { \key c \alhizazkar c'4 db' e' f' g' ab' b' }

give's Sahnaz in D.

Sahnaz is therefore redundant. I could delete the entry but commenting
it out maybe a better, long term option for others who may expect to be
able to use 'Sahnaz' as a Makam name.