Issue 5551/1: better indexing support
In TeX output mode we unify `@findex', `@cindex', and `@vindex'
entries in the concept index; this is done with the `@syncodeindex'
lines in macro `@lilyTitlePage'.  We also set
`txiindexbackslashignore' so that `\foo' gets sorted as `foo'.
However, the latter directive has the side effect that `\foo' and
`foo' are treated as equal by the index sorting program.  We don't
want that: It is important for us that entries for both `\foo' and
`foo' appear in the unified index.

For this reason, we make macro `@funindex' use `@sortas' to set a
special sorting key, and we redefine texinfo's internal macro
`\indexwritesortas' to swallow backslashes.  We also modify
`@cindex' to insert another sorting key to ensure that `\foo'
immediately comes after `foo', followed by entries like `foo bar'.

(Note that HTML output is not improved by this change since the
ancient texi2html version we have to use doesn't support `@sortas'.)

Also add option `-8bit' to xe(la)tex to get real 8bit characters in
auxiliary index files.

Also emit `@vindex' entries as slanted; explain this in
`notation.tely'.