Issue 5555/3: TimeScaledMusic should not be music-wrapper-music
The original title of this issue was "\tweak no longer works with \tuplet".

The trigger of the problem is issue 5302 which lets tweaks bypass
expressions marked as music-wrapper-music.  \tuplet creates
TimeScaledMusic expressions which take tweaks for applying to the
created TupletBracket and TupletNumber, so this kind of passing on
does not seem appropriate.  Removing the music-wrapper-music type has
several other consequences for tuplets:

* Music functions returning or music variables containing a
  single-note \tuplet expression can no longer be used inside of a
  chord (where their duration would be overriden anyway).

* Post events can no longer be attached to a \tuplet expression as a
  whole.  Previously, this let them migrate to the last note of the
  tuplet expression.

While the former appears to be a non-issue because of being obscure
and basically useless, the latter is addressed with a separate patch.