Issue 5568: make build output terse by default
The goal is to show progress without burying warnings and errors in
the noise of long commands.

Options VERBOSE=1 and SILENT=1 make the output more like the usual
"make" and "make -s" (respectively).

The verbosity level affects the options that are passed to some
commands, but there is room for improvement in this area.