Issue 5578: add a button to flip between old and new regtest images
If JavaScript is enabled, a "Flip" button appears under the right-hand
image.  While the button is held down, the right-hand image is
replaced by the left-hand image.  Flipping back and forth can make
small differences more obvious than static highlighting does.

The output-distance script used to use HTML tags that are unsupported
in HTML5, such as <tt> and <font>.  The new approach uses CSS for
general styling as well as for the flipping feature.

The new style is not exactly the same as before, but should look
familiar.  The style sheet also includes a dark color scheme which
newer browsers may select based on user preference.

I found the program "tidy" to be somewhat useful for catching problems
like empty or unclosed tags in the generated HTML, and I added it to
the build as an option.  http://validator.w3.org caught a few more
problems, but tidy is fast and I didn't see it raise any incorrect
objections, so it seems beneficial to use it when it is available.