Issue 5578: add a button to flip between old and new regtest images
This is a resubmission of 911788f173eb58438fc9c850a005638d053b8bba
minus the use of "tidy" to check the generated HTML.

If JavaScript is enabled, a "Flip" button appears under the right-hand
image.  While the button is held down, the right-hand image is
replaced by the left-hand image.  Flipping back and forth can make
small differences more obvious than static highlighting does.

The output-distance script used to use HTML tags that are unsupported
in HTML5, such as <tt> and <font>.  The new approach uses CSS for
general styling as well as for the flipping feature.

The new style is not exactly the same as before, but should look
familiar.  The style sheet also includes a dark color scheme which
newer browsers may select based on user preference.