Issue 5593/2: optionally use "tidy" to check generated HTML
I found the program "tidy" to be somewhat useful for catching problems
like empty or unclosed tags in the HTML generated by output-distance,
so I added it to the build as an option.  http://validator.w3.org
caught a few more problems, but tidy is fast and I didn't see it raise
any incorrect objections, so it seems beneficial to use it when it is
available.