Issue 5593/1: use "false" for missing programs
STEPMAKE_OPTIONAL_REQUIRED sets "false" for a missing program, whether
required or optional.

I removed a comment suggesting to abort immediately when a required
program is missing; that would be less helpful than the current behavior
of reporting all missing programs when the script is done.

Thanks to Jonas Hahnfeld <hahnjo@hahnjo.de> for correcting a
misunderstanding I had about this code, leading to this improvement.