Issue 5597: musicxml: Use method to get number of Lyric
Commit edb29ef3a ("Correct an inconsistency in the Lyric class")
renamed the method from get_number() to number(), which is shadowed
by the attribute itself and not used in the current code. Using a
method has the advantage that the code can cast to int() in a single
place which is needed for Python 3. Additionally, the code handles
the case if there is no 'number' attribute (should this ever happen).