Issue 5621/2: explicitly align rehearsal mark in pedal-bracket test
The mark is obviously intended to test the extent of the pedal bracket.
It is not clear whether it matters that the mark extends to the left as
well as to the right, but to be safe, explicitly center it so that
future changes to the default alignment do not affect this test.