Issue 5648/3: Stop copying FlexLexer.h
Leave FlexLexer.h where it is, rather than copying it to $(outdir).

As we enable more warnings, we will not want to hear about problems in
this file.  g++ automatically suppresses warnings for headers found in
system include directories (specified with -isystem), so the simple
solution is to leave it where it is.