Issue 5663/1: Use codecs.open() to decode as utf-8
This is in preparation for Python 3.5 where the default encoding
depends on the value of the LANG environment variable. As far as
I can tell, this was changed later on and at least Python 3.7 and
version 3.8 always default to 'utf-8' on Linux. As I'm proposing to
make Python 3.5 the required minimum, we can't rely on this and need
to force 'utf-8' when reading files that could contain Unicode.