Issue 5689: Refactor enforcement of a single Score context
Context::find_create_context () and Context::create_unique_context ()
no longer require special cases to avoid ascending to the Global context
or to create a Score themselves.  Instead, the method that they each
call to create any new context, Context::create_hierarchy (), avoids
creating multiple Scores.

This is still hard-coded in C++ to happen for Score only, but it would
be straightforward to add a context attribute that could be used in ly
code like \accepts and \defaultchild.