Issue 5665/1: Drop -dtrace-scheme-coverage
As the initial commit dee82b1d0b of scm/coverage.scm notes, this
module requires support for memoize-symbol. This was indeed added
for Guile 1.9.0, but again removed in version 1.9.13. That means
there never was a stable version of Guile that worked with this.

When building / running with current Guile 1.8.8, the executable
crashes because the interpreter raise()s from within a critical
section and (at least my) libc kills the process:
 $ guile1.8 -c "(trap-enable 'memoize-symbol)"
throw from within critical section.
Aborted (core dumped)

(With later releases of Guile, there's at least an error message:
 $ guile2.0 -c "(trap-enable 'memoize-symbol)"
[...]
ERROR: In procedure memoize-variable-access!:
ERROR: Unbound variable: trap-enable

This is because the mentioned (unstable) release 1.9.13 completely
removed not only memoize-symbol, but the whole trap-* interface.)