Grow heap aggressively during music interpretation
During the interpretation stage, we build up a large set of
Grobs. Without tuning, GC is ineffective, because the Grobs are live
data and can't be collected.

To remedy this, if we see an unsuccessful GC step in the
Score_engraver we increase the heap size explicitly

Total GC overhead for mozart-hrn-3:
  without initial heap size: 1.7s
  with initial heap size set:  ~1.2s
  with heap scaling: ~0.25s

Change-Id: I8ab8c845afbda0d767f506dda45c80fcbaeee950