Add Unicode filename support for Windows 10 1903+
On Windows 10 1903 and above, -A APIs support UTF-8 encodings.
This commit enables it by adding an application manifest.
https://docs.microsoft.com/en-us/windows/uwp/design/globalizing/use-utf8-code-page

If you use older Windows,
it has no effect and the behavior is the same as before.

For Cygwin environments, UTF-8 can be used originally,
so this commit doesn't add the application manifest.