Issue 5734: Fix convert-ly with Python 3
The error was "'str' object has no attribute 'decode'", so it already
is decoded. The encoding seems to also be correct, as filenames with
Unicode characters work just fine.
It's worrying that none of the usual testing commands seem to invoke
the script with a real input file, only the output of --help is used.