GUILE2: Scale GC heap with the number of smobs
On every GC cycle, increase the heap so we have at least 2000 bytes
per non-simple Smob. This increases the size of the heap relative to
the live set of data.

The number 2000 is based on benchmarking the Robert Carver's "Missa
Dum sacrum mysterium" Edited by Vaughan McAlley (2013), a 104 page
choir score.

By setting GC_INITIAL_HEAP_SIZE and the max heap size, we can see how
GC time and heap size relate.  On a Lenovo T460p (i5 6440HQ 2.6Ghz),
this yields the following timings:

* 2G heap: 49.4s (1.8s GC)
* 1G heap: 51.7s (3.3s GC)
* 800M heap: 54.6s (5.8s GC)
* 500M heap: 68.2s (17.9s GC)

We see that 800M - 1G is a sweet spot, where around 10% of time is
spent on GC. By instrumenting the code, we see that there are about
2000 bytes per smob in this case.  Running the same score with
automatic scaling to 2000 bytes per smob, we get a final heap size of
1.2G, and a GC time of 3.4s.

The reason this is necessary, is that BDWGC has a special case for
objects with finalizers

  https://github.com/ivmai/bdwgc/blob/v8.0.4/alloc.c#L1435

where it decides to not expand the heap if there are more than 500
objects with finalizers outstanding.  Since smobs with free functions
are implemented with finalizers, we fall into this case often.

The symptom of this problem is that running with GC_PRINT_STATS=1 will
print

  In-use heap: 85% (370824 KiB pointers + 60065 KiB other)
  In-use heap: 85% (370725 KiB pointers + 59737 KiB other)
  ..


https://sourceforge.net/p/testlilyissues/issues/5718
http://codereview.appspot.com/561390043