Issue 5779/1: [SVG] Permit 'e' to appear in SVG font glyph paths.
The definition of 'glyph-path-regexp' will only match a <path d="...">
element if the contents of the 'd' attribute do not include the letter
e. But the SVG spec permits coordinates to be given in C-style
exponential notation such as 9.8e-5. If a glyph in an SVG font uses
that ability, extract-glyph will fail, and the glyph will be silently
omitted from the output file.

(cherry picked from commit e45b5186b25e4ec4077621d917aaa014be342535)