Accept GUILE 2 without extra configure options
A run of the regression tests at "5f4cf7bd38 - Parse inline scheme
using per-expression port" shows the following issues compared to
GUILE 1.8

* changed interleaving of error and progress messages.

* floating point rounding differences.

* an encoding error in song-basic-nonenglish.ly

As of a19aed1 "GUILE2: Scale GC heap with the number of smobs",
performance for large scores with GUILE 2.2 is just ~25% slower on
large scores.

The regression test (make test-baseline) is 67% slower. This can
partly be explained because of the slower startup: until we can
support pre-compiled files, we incur a 1.5 sec extra delay for parsing
.scm files.

https://sourceforge.net/p/testlilyissues/issues/5780
http://codereview.appspot.com/569390043