GUILE2: Do not run GC after processing every file.
When processing multiple files, we ran GC between the files.  For a
conservative garbage collector this yields a more accurate scan,
because then our stack is small. We never measured the impact of this
idea.

This change was originally motivated by the dead-object detection for
parsing related objects, which needs a GC run to mark the supposedly
dead object as live.

As a result, on every GC, we paiud the cost of tracing through the
live set, ie. fixed data that spans multiple files, and by forcing GC
calls, we traced that data more often than necessary.

We get CPU time savings of ~20 % by not forcing GC.

For GUILE 1.8, the following timings for running "make test-baseline"
on my development branch

force GC on every file

  real 3m23.886s
  user 8m6.921s

let the GC decide when to collect

  real  3m10.523s
  user  6m36.925s

For GUILE 2.2, the difference is similar.

This patch only disables the GC for GUILE 2.x. For GUILE 1.8 we still
leave the object-should-be-dead detection enabled.

https://sourceforge.net/p/testlilyissues/issues/5797
http://codereview.appspot.com/579330043