Issue 5822/2: Replace HTML size comments with human-readable size
This calculates download sizes as the last step of HTML generation and
puts them into the HTML files for reference.

The work is done by scripts/build/fix-docsize.sh .

It does that by replacing the size-requesting HTML comments in the
files given on the command line by the respective human-readable size.

scripts/build/fix-docsize.sh is run in the respective directories as
the final step of generating the HTML versions.

This script uses ls -sh in order to get a human-readable size spec.