Add a FS lock to lilypond-book
This simplifies the build infrastructure, because it obviates Makefile
hacks to force a single lilypond-book processes during the build

This uses a fcntl-based lock, which is managed by the kernel. This
means there is no chance of the locks going stale (except on NFS)

Remove limitations on parallelism in build system.

https://sourceforge.net/p/testlilyissues/issues/5787
http://codereview.appspot.com/555360043