Issue 5828: Fix Windows resource to enable manifest
A manifest file `lilypond.manifest` is specified
in resource file `lilypond.rc` for the Windows executables.
To specify the manifest, the definitions
`CREATEPROCESS_MANIFEST_RESOURCE_ID` and `RT_MANIFEST`
are required in the resource file.
However, the manifest was not activated
because the resource file did not include `winuser.h`
that contained the required definitions.
Threfore, Unicode filenames could not be used
because the manifest was not activated.

This commit adds the include file `winuser.h` to the resoruce file
and activates the manifest.
This allows LilyPond for Windows to use Unicode filenames.