Issue 5833/2: Add workaround for MinGW UTF-8 stat issue
When UTF-8 process locale is active,
LilyPond could not find some Unicode filenames.

`stat ()` in msvcrt.dll returns `-1` (i.e. error) with some UTF-8
filenames even if process code page is UTF-8 and the file exists.
So we get the process code page with `GetACP ()` and, if it is UTF-8,
convert the filename to wide strings and use `_wstat ()`.

If Universal CRT (UCRT, newer than msvcrt.dll) is used
(i.e. `_UCRT` is defined), no such workaround is needed.