flower: Delete library.a before creating
Apparently i686-apple-darwin8-ranlib in GUB produces archives that
become 'malformed' if further processed by 'ar'. I'm not sure _why_
this currently happens in GUB, but make tries to rebuild the archive
just before linking lilypond. The issue is that ar tries to _update_
the archive if it already exists instead of recreating from scratch.

As a solution just manually delete library.a (if it happens to exist)
as part of the rule to create it. Also drop the 'u' modifier to only
update newer files - it was ignored anyway with the new deterministic
default 'D'.