weblinks: Fix encoding of link texts
When switching to Python 3, I added the call of .encode('utf-8') to
fix the build with Python 3.5. This converts all link texts into
binary encoding, enclosed by b'...'. The correct fix is to force
utf-8 encoding for sys.stdout.
See e0c78a4c71 ("Use codecs.open() to decode as utf-8") for more
details on this issue and how the default changes with Python 3.7.