Copy alist instead of deep copy on Grob clone
On the MSDM score, ly_deep_copy takes 1.58% of runtime according a
gprof. Removing it shows a consistent 1.4% improvement in user
time.

The deep copy was introduced together with the separation of object
properties and value properties. This is because the Grob substition
did a recursive rewrite of the entire mutable alist, and the
equivalent for value (ie. non-object) properties is to do a deep copy.

The deep copy does not really solve a problem, because the reader
calling get_property cannot distinguish grob-local values from
immutable, shared values. The sort of modification that a deep copy
protects against would modify the shared data in this case, leading to
bugs.

https://sourceforge.net/p/testlilyissues/issues/5906
http://codereview.appspot.com/561640045