Issue 5917: ly:performance-headers and \midi { after-writing }
Change ly:performance-header into ly:performance-headers, returning a
list of visible header modules from innermost to outermost.  Before,
only the innermost header was provided, so the values of variables
defined in outer headers were not accessible.

If the \midi block defines a function called after-writing, LilyPond
calls it after writing the MIDI file.  Apart from its use in regression-
testing ly:performance-headers, it offers a way for an advanced user to
dump metadata associated with each performance.  For example, the
function could open a file and dump all header variables in YAML format
for use in a scripted workflow where the information ends up in ID3 tags
in an mp3 file.