Issue 5921: Documentation: Remove deprecated skeleton-update
It's been like that since commit 858e506c2a in 2010, ten years ago.
Also remove target all-translations-update which cannot have worked
since the same time and update some CG pages mentioning the targets
and the underlying script texi-skeleton-update.py.

Rewrite the section about 'Adding a Texinfo manual' because it was
out-of-date with respect to the build system: Directories with the
included .itely files are not listed in SUBDIRS and have no
GNUmakefile that would need to be copied.