Issue #1204: Document, and add regtest for, external fonts.
Add a few details in NR 1.8.3.1 "Fonts explained", and
a regtest with base64-embedded minimal OpenType fonts.
(I didn’t want to use musical symbols since the whole
point is that these glyphs are not to be confused with
whatever’s installed on the OS.)
The regtest has been tested with all major versions
since 2.12, and current master with and without Guile2.