Issue 5930: Make dblatex an optional dependency
Commit cfa910f1c1 in 2012 made it required unless disabling the doc
build. This is an overapproximation since dblatex is only required
for compiling the DocBook result of running lilypond-book on the file
input/regression/lilypond-book/suffix-lyxml.lyxml. If unavailable the
rules in make/lilypond-book-*.make have checks to only produce the
intermediate .xml file, without running dblatex to produce a .pdf.