Issue5947/1: Add `-dfont-ps-resdir` option to embed fonts to PDFs later
A similar purpose option, `-dfont-export-dir` exports fonts
as individual PostScript files for font embedding later.
However it cannot embed CID fonts to PDFs with Ghostscript 9.26 and later.

This `-dfont-ps-resdir` option creates
a subset of PostScript resource directory
in the directory specified by the option.
The directory contains symbolic links to font files
that Ghostscript can use to embed the fonts to PDFs later.
This method can embed CID fonts, even with Ghostscript 9.26 and later.

This `-dfont-ps-resdir` option skips TrueType fonts
because garbled characters occur same as `-dfont-export-dir` option.
Unlike `-dfont-export-dir` option,
it skips OTC (OpenType Collection/CFF) font files
because Ghostscript cannot handle them directly.

Simple usage:

```
$ lilypond -dfont-ps-resdir=resdir -dgs-never-embed-fonts foo.ly
$ gs -q -dBATCH -dNOPAUSE -sDEVICE=pdfwrite \
     -I resdir -I resdir/Font \
     -sOutputFile=foo.embedded.pdf foo.pdf
```

Note:
It is better not to specify the directory
that contains the name `Resource` because it has a special meaning
when specified with the -I option for Ghostscript.