Add CI configuration
This adds automatic testing, very similar to the manual process
used so far. The image is based on Ubuntu 16.04 with Python 3.5
which makes sure that everything continues to work with the
oldest supported version.

As a design choice split the pipeline into three stages:
 - build to configure and 'make' LilyPond,
 - test to call 'make test', and
 - documentation to call 'make doc'.
The rationale is that we might want to add more configurations later
on, for example testing a newer version of Ubuntu or in-tree builds.
As each stage waits for all jobs from the previous stage, this can
avoid needless 'test' jobs if a 'build' fails.

To improve performance, the configuration disables debug symbols and
uses the Ghostscript API for faster documentation.