Issue 5129: Bad horizontal spacing when \override LyricText #'X-offset
The Problem
===========
The change in issue #2462 allowed the ideal distance of a spring to be less
than its min_distance.  Most of the time this is desirable. However in the
circumstance where one spring in a rod is much larger than all the others, the
return value of rod_force results in rods that are much longer than their
intended length.  This happens because rod_force (through range_ideal_len)
bases its calculations on the ideal distance of each spring, *not* the
min_distance.  And since issue 2462 this value can now be much smaller,
therefore rod_force returns larger values, sometimes much larger.

The net effect of this is that in lines containing a single very large spring -
one that has a small ideal distance, but a large min_distance (e.g. a note with
one very long lyric syllable) - it is often not even able to fit one bar in the
line, even though it should be (and was before 2462). It causes bars to exceed
line-width when not necessary or even run off the page.

The Solution
============
Replace rod_force with a heuristic function that tries not to overestimate the
force required to produce the desired distance. Add a regression test.

(Reverting 2462 is undesirable because it improves spacing most of the time and
the scenario that triggers the above problem is relatively rare. Other simple
solutions that were tested invariably either made things worse or fixed the
issue, but reintroduced the poor spacing from before 2462.)

The Results
===========
This both keeps the benefits of issue 2462 while restoring the previous
behaviour in the problematic scenario described above.

This increases LilyPond's ability to compress lines. It mainly affects scores
with one or more of the following properties:
- are ragged right
- fit on one line
- have a small line-width

For example, it fixes a problem with the regression test non-empty-text.ly:
after this change LilyPond is able to fit the second system into the specified
line-width (3cm), whereas before this change it was not able to do so.

Closes #5129