Music_iterator: derive `ok` from `pending_moment` and `run_always`
`pending_moment` returns an upper bound on the next moment to be
processed, or +infinity if there is no bound.

`run_always` could have a better name: it can change value during the
life of the iterator.  When true, it means that `process` should be
called at moments earlier than `pending_moment`.

An iterator is `ok` if `pending_moment` is finite or it wants to
`run_always` regardless.