Lazy_skyline_pair: fix right to left segments with thickness
Commit 267e894 ("Thread skyline construction through stencil
interpretation") simplified line thickness handling, by transforming
the line segment, and then widening it

The approach to widen the segment only works if p1 is left of p2.  If
p1 is on the right, the segment will be shortened, effectively making
it the curve equivalent to a stippled line. To remedy this, swap the
points when necessary.

This fixes sawtooth skylines in input/regression/skyline-grob-rotation.ly