Use eq? for alist lookups with symbols
Inspection of

  perf record -g  lilypond -I carver/ MSDM-reduced
  perf report --call-graph --stdio

attributes 3.95% of runtime to scm_equal_p and ~5.52% to
scm_string_eq. Inspection with GDB suggests that the majority of these
calls come from ly_chain_assoc_get: every markup or text uses multiple
lookups using ly_chain_assoc_get to determine what font to use.

GUILE could shortcut scm_is_equal if one of the keys is a symbol, but
it doesn't, so we do it ourselves instead: this commit introduces
ly_assoc, which shortcuts to scm_assq if the key is a symbol or an
immediate

Timing data:

  7d5e7e6eae - Use eq? for alist lookups with symbols
    baseline 43319677c5 - Cache the name => index lookup in Open_type_font
    args [input/regression//mozart-hrn-3]
    mem delta: -103544.000000 (stddev 66.090847 82.397411 n=3)
    mem delta: -23.018371 % (7d5e7e6eae is leaner)
    time delta: -0.120000 (stddev 0.010000 0.005774 n=3)
    time delta: -3.582090 % (7d5e7e6eae is faster)

  7d5e7e6eae - Use eq? for alist lookups with symbols
    baseline 43319677c5 - Cache the name => index lookup in Open_type_font
    args [-I carver MSDM-reduced]
    mem delta: -102080.000000 (stddev 98.332768 62.780039 n=3)
    mem delta: -12.871402 % (7d5e7e6eae is leaner)
    time delta: -1.820000 (stddev 0.032146 0.020817 n=3)
    time delta: -7.871972 % (7d5e7e6eae is faster)